First pass at feedback
Seemed like this was the easiest way to leave comments on the whole
project. To be honest, I haven't worked in Rust for a few years now, and
was never the Rust style expert. That being said, overall I think you're
doing a great job taking advantage of the features of the language. I
didn't have any major design comments up front, I left some initial
thoughts in comments in the code. I'll think more about it and see if
anything else occurs to me!